refactor remove functions
