+++ lib/methods/export.js
@@ -367,15 +367,6 @@ module.exports = self => {
       }
     },
 
-    async remove(filepath) {
-      console.info(`[export] removing ${filepath}`);
-      try {
-        await fsp.unlink(filepath);
-      } catch (error) {
-        self.apos.util.error(error);
-      }
-    },
-
     // Report is available for 10 minutes by default
     removeFromUploadFs(downloadPath, expiration) {
       const ms = expiration || 1000 * 60 * 10;
