+++ lib/methods/import.js
@@ -132,7 +132,7 @@ module.exports = self => {
           .dismiss(req, notificationId, 2000)
           .catch(self.apos.util.error);
 
-        await self.removeExport(exportPath);
+        await self.remove(exportPath);
         return;
       }
 
@@ -531,27 +531,7 @@ module.exports = self => {
         self.apos.notification.dismiss(req, notificationId, 2000).catch(self.apos.util.error);
       }
 
-      await self.removeExport(exportPath);
-    },
-
-    async removeExport(filepath) {
-      try {
-        const stat = await fsp.lstat(filepath);
-        if (stat.isDirectory()) {
-          await fsp.rm(filepath, {
-            recursive: true,
-            force: true
-          });
-        } else {
-          await fsp.unlink(filepath);
-        }
-        console.info(`[import] export path ${filepath} has been removed`);
-      } catch (err) {
-        console.trace();
-        self.apos.util.error(
-          `Error while trying to remove the file or folder: ${filepath}. You might want to remove it yourself.`
-        );
-      }
+      await self.remove(exportPath);
     }
   };
 };
