+++ lib/methods/index.js
@@ -1,3 +1,4 @@
+const fsp = require('node:fs/promises');
 const importMethods = require('./import');
 const exportMethods = require('./export');
 
@@ -34,6 +35,27 @@ module.exports = self => {
 
       return true;
     },
+
+    async remove(filepath) {
+      try {
+        const stat = await fsp.lstat(filepath);
+        if (stat.isDirectory()) {
+          await fsp.rm(filepath, {
+            recursive: true,
+            force: true
+          });
+        } else {
+          await fsp.unlink(filepath);
+        }
+        console.info(`removed: ${filepath}`);
+      } catch (err) {
+        console.trace();
+        self.apos.util.error(
+          `Error while trying to remove the file or folder: ${filepath}. You might want to remove it yourself.`
+        );
+      }
+    },
+
     ...importMethods(self),
     ...exportMethods(self)
   };
